refactor(Domain): 도메인 패키지 생성 및 해당 클래스 이동
요구 사항에서, Lotto의 패키지 변경이 가능하므로 해당 클래스를 도메인으로 모아 정리함
이에 따른 테스트 디렉토리도 변경
