+++ src/main/java/lotto/domain/LottoBall.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.exception.ArgumentRangeOutOfBoundsException;
